✨ feat: GatheringChallengeCard 분리
